+++ src/main/kotlin/com/github/dweindl/intellijAntimony/AntimonyBundle.kt
@@ -5,9 +5,9 @@ import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.AntimonyBundle"
 
-object MyBundle : DynamicBundle(BUNDLE) {
+object AntimonyBundle : DynamicBundle(BUNDLE) {
 
     @JvmStatic
     fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =